Simplify Newton global import defaults
